comgt-ncm: Add driver dependencies again
In the commit  "comgt-ncm: Fix NCM protocol" the dependencies to vendor NCM drivers were removed, because

> comgt-ncm should not depend on the USB-serial-related kernel modules, as the cdc-wdm
> control device works without them. There is also no need to depend on
> kmod-huawei-cdc-ncm, since other manufacturers (like Sony-Ericsson
> and Samsung) which use other kernel modules should also be supported.

This does not make sense from a user-perspective as installing comgt-ncm (or luci-proto-ncm) should install all needed dependencies
for using such a device. Furthermore depending on kmod-huawei-cdc-ncm does not mean that Sony-Ericsson and Samsung
devices can't be supported. By the way it seems that Sony-Ericsson and Samsung devices never used NCM, but act as serial modems.
Thus this commit adds the dependencies again.